admin/backend/requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pgvector==0.2.4

# Redis
redis==5.0.1

# AI & ML
google-generativeai==0.3.2
anthropic==0.8.1
openai==1.7.2
pillow==10.2.0
numpy==1.26.3

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0
playwright==1.49.1

# Google Cloud
google-cloud-texttospeech==2.14.2
google-cloud-translate==3.14.0
google-cloud-storage==2.14.0
google-api-python-client==2.190.0
google-auth-oauthlib==1.2.0

# Utils
python-dotenv==1.0.0
loguru==0.7.2
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3
APScheduler==3.10.4

# Celery
celery==5.3.4
flower==2.0.1

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
isort==5.13.2
